feat: permit only user owner to delete or modify their account

+++ src/main/java/com/hui/plateform/hiuplateformeapi/controller/UserController.java
@@ -4,6 +4,8 @@ import com.hui.plateform.hiuplateformeapi.Service.UserService;
 import com.hui.plateform.hiuplateformeapi.entity.User;
 import lombok.AllArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -28,13 +30,30 @@ public class UserController {
     }
 
     @PutMapping("/user/{id}")
-    public User updateUser(@PathVariable("id") String id, @RequestBody User user){
-        user.setId(id);
-        return userService.createUpdateUser(user);
+    public User updateUser(@PathVariable("id") String id, @RequestBody User user, Authentication authentication){
+        String currentUserId = getCurrentUserId(authentication);
+        if (id.equals(user.getId()) && id.equals(currentUserId)) {
+            userService.createUpdateUser(user);
+        }
+        else{
+            throw new IllegalStateException("Cannot update another user");
+        }
     }
 
     @DeleteMapping("/user/{id}")
-    public void deleteUserById(@PathVariable("id") String id){
-        userService.deleteUser(id);
+    public void deleteUserById(@PathVariable("id") String id, Authentication authentication){
+        String currentUserId = getCurrentUserId(authentication);
+        if (!id.equals(currentUserId)) {
+            userService.deleteUser(id);
+        }
+        else{
+            throw new IllegalStateException("Cannot update another user");
+        }
+    }
+    private String getCurrentUserId(Authentication authentication) throws IllegalStateException {
+        if (authentication == null) {
+            throw new IllegalStateException("Authentication required.");
+        }
+        return ((User) authentication.getPrincipal()).getId();
     }
 }
